Revamp type generation to support forward references
- Add cross-referencing test protos
- Implement pass 1
  * create Pass 1 functions
  * start flattened generation
  * flattened name mangling
  * move flattened name generation into the resolver
  * recursive type generation
  * detecting circular dependency, error
- Implement pass 2
  * special care to enums
- Refactor, split up common parts from both passes into separate files

Also:

- Add generated proto.js file to the coverage ignores
  * Keep the res file covered, but ignore the js generated one.
