+++ packages/bs-protobuf/src/compiler/gen-types/emit-constants.js
@@ -1,0 +1,1 @@
+export const bsProtobufPackage = "ReprotoBsProtobuf";
