+++ packages/bs-protobuf/test/proto/module_test_a1_3.proto
@@ -1,0 +1,8 @@
+syntax = "proto3";
+
+package moduleTestA;
+
+message TestMessageA1 {
+  optional string stringField = 1;
+}
+
