+++ packages/bs-protobuf/test/proto/module_test_a2_3.proto
@@ -1,0 +1,10 @@
+syntax = "proto3";
+
+package moduleTestA;
+
+import "module_test_b1_3.proto";
+
+message TestMessageA2 {
+  optional moduleTestB.TestMessageB1 messageField = 1;
+}
+
