+++ packages/bs-protobuf/test/proto/module_test_b1_3.proto
@@ -1,0 +1,10 @@
+syntax = "proto3";
+
+package moduleTestB;
+
+import "module_test_a1_3.proto";
+
+message TestMessageB1 {
+  optional moduleTestA.TestMessageA1 messageField = 1;
+}
+
