+++ packages/bs-protobuf/test/proto/module_test_b2_3.proto
@@ -1,0 +1,8 @@
+syntax = "proto3";
+
+package moduleTestB;
+
+message TestMessageB2 {
+  optional string stringField = 1;
+}
+
