+++ packages/bs-protobuf/test/proto/type_test_3.proto
@@ -52,4 +52,5 @@ message Typeful {
   map<int64, string> mapInt64StringField=23;
   optional bool boolField=24;
   map<bool, string> mapBoolStringField=25;
+  google.protobuf.Any anyField=26;
 }
